+++ frontend/app/src/modules/history/management/forms/form-guards.ts
@@ -8,7 +8,7 @@ import type {
   SwapEvent,
 } from '@/types/history/events/schemas';
 import { HistoryEventEntryType } from '@rotki/common';
-import { isEvmEvent, isSolanaEvent, isSolanaSwapEvent } from '@/utils/history/events';
+import { isEvmEvent, isEvmSwapEvent, isSolanaEvent, isSolanaSwapEvent } from '@/utils/history/events';
 
 export function isGroupEditableHistoryEvent(event: HistoryEvent): event is GroupEditableHistoryEvents {
   return event.entryType === HistoryEventEntryType.ASSET_MOVEMENT_EVENT
@@ -17,10 +17,6 @@ export function isGroupEditableHistoryEvent(event: HistoryEvent): event is Group
 
 export function isSwapEvent(event: HistoryEvent): event is SwapEvent {
   return event.entryType === HistoryEventEntryType.SWAP_EVENT;
-}
-
-export function isEvmSwapEvent(event: HistoryEvent): event is EvmSwapEvent {
-  return event.entryType === HistoryEventEntryType.EVM_SWAP_EVENT;
 }
 
 export const EVM_EVENTS = [